requirements.txt
# Core dependencies
python-telegram-bot>=13.0
pymongo>=4.0.0
python-dotenv>=0.19.0
requests>=2.27.0
google-generativeai>=0.3.0

# Data processing and visualization
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Web dashboard (optional for future extension)
flask>=2.0.0
flask-mongoengine>=1.0.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.62.0
loguru>=0.6.0